requirements.txt
async-timeout==3.0.1
atomicwrites==1.4.0
backcall==0.2.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
biopython==1.79
boto==2.49.0
Bottleneck==1.3.2
celery==5.2.1
certifi==2021.10.8
cffi==1.14.0
cftime==1.2.1
chardet==3.0.4
configparser==5.1.0
contextlib2==0.6.0
coverage==5.3
cycler==0.10.0
Cython==0.29.17
distributed==2.5.2
dj-database-url==0.5.0
Django==3.2.7
django-computedfields==0.1.5
django-cors==0.1
django-cors-headers==3.10.0
django-jsonfield==1.4.1
djangorestframework==3.12.4
future==0.18.2
get_version==2.1
gunicorn==20.1.0
install==1.3.4
lazy-object-proxy==1.4.3
legacy-api-wrap==1.2
libarchive-c==2.9
locket==0.2.0
lxml==4.5.1
Markdown==3.3.1
msgpack==1.0.0
multidict==5.1.0
multipledispatch==0.6.0
olefile==0.46
openpyxl==3.0.0
packaging==20.9
partd==1.0.0
path.py==12.0.1
pathtools==0.1.2
pipreqs==0.4.11
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psycopg2-binary==2.9.2
rq==1.10.0
sinfo==0.3.1
six==1.15.0
Sphinx==2.2.0
stdlib-list==0.8.0
tables==3.6.1
toolz==0.10.0
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==5.1.0
widgetsnbextension==3.5.1
wrapt==1.11.2
yarg==0.1.9
yarl==1.6.3
zict==1.0.0
zipp==0.6.0
zope.event==4.5.0
